feat(custom-connector): externalize the Custom Connector wizard and management table
Moves CustomConnectorWizard and the "Custom Connectors" API-key table out of
CrawlersPage.jsx into tools/crawlers/custom-connector/, matching the existing
CSV/Omada/midPoint/Entra ID plugin pattern. Introduces a new optional
Table.jsx plugin file for crawler types that aren't CrawlerConfigs rows (a
standalone list of N items rather than a per-config card) — CrawlersPage.jsx
auto-discovers it and remounts it via a tableRefreshKey bump whenever any
wizard completes, since the table now owns its data privately.

Also removes custom-connector from the crawler-manifest CI job's
PENDING_MIGRATION exemption list, so the "no crawler-specific drift into
app/ui" check now actually covers this type.
