+++ .github/workflows/pr.yml
@@ -361,7 +361,7 @@ jobs:
 
           # Crawlers pending migration to the manifest-based plugin system.
           # Remove a type from here once its CrawlerMeta.js is in place.
-          PENDING_MIGRATION="demo custom-connector"
+          PENDING_MIGRATION="demo"
 
           FAIL=0
