+++ tools/crawlers/custom-connector/CrawlerMeta.js
@@ -1,0 +1,5 @@
+export default {
+  id: 'custom-connector',
+  name: 'Custom Connector',
+  description: 'Build your own crawler using the Ingest API — register an API key, download the OpenAPI spec, start pushing data',
+};
